added error handler

+++ controllers/categoryController.js
@@ -6,12 +6,25 @@ async function getAllCategories(req, res) {
   res.json(categoryDtos)
 }
 
-async function postNewCategory(req, res) {
+async function postNewCategory(req, res, next) {
+  if (req.auth.accountType != 'admin') {
+    let error = new Error('Unauthorized, not and admin')
+    error.status = 401
+    next(error)
+    return
+  }
   await categoryService.addCategory(req.body)
   res.sendStatus(201)
 }
 
-async function deleteCategoryById(req, res) {
+async function deleteCategoryById(req, res, next) {
+  if (req.auth.accountType != 'admin') {
+    let error = new Error('Unauthorized, not and admin')
+    error.status = 401
+    console.log(error)
+    next(error)
+    return
+  }
   await categoryService.deleteCategoryById(req.params.categoryId)
 }
