+++ services/postService.js
@@ -1,4 +1,5 @@
 const postRepository = require('../repositories/postRepository')
+const userRepository = require('../repositories/userRepository')
 const { postDto } = require('../mappers/postMapper')
 const postModel = require('../models/postModel')
 
@@ -8,7 +9,13 @@ async function getPostPageByThreadId(threadId, options) {
   return page
 }
 
-async function createNewPost(postDto, user) {
+async function createNewPost(postDto, username) {
+  const user = await userRepository.getUserByUsername(username)
+  if (!user) {
+    let error = new Error('User not found for name: ' + username)
+    error.status = 400
+    throw error
+  }
   let post = {
     contents: postDto.contents,
     author: user._id,
@@ -23,7 +30,23 @@ async function getPostById(postId) {
   return postDto(post)
 }
 
-async function deletePostById(postId) {
+async function deletePostById(postId, username) {
+  const user = await userRepository.getUserByUsername(username)
+  if (!user) {
+    let error = new Error('User not found for name: ' + username)
+    error.status = 400
+    throw error
+  }
+  
+  const foundPost = await postRepository.getPostById(postId)
+
+  //check if retrieved user is an admin or the author 
+  if (user._id != foundPost.author && user.accountType != 'admin') {
+    let error = new Error('Not authorized to delete the post')
+    error.status = 401
+    throw error
+  }
+
   await postRepository.removePostById(postId)
 }
